Add files via upload
Field phase
Four Acoustic Doppler Current Profiler (ADCP) were installed in the northern part of the PNSAV in front of the port of Veracruz. The four profilers were deployed on July 22, 2008, and were subsequently collected on October 2, 2009, a period that covers the season from autumn to spring and covers at least one period of the passage of high- and low-pressure systems in the western coastal area of the Gulf of Mexico.
The position of the four profilers was: two Nortek Aquadop 400 kHz ADCPs on the leeward side of Blanquilla and Anegada Adentro reefs, and two 1 MHz Nortek Aquadop ADCPs on the south side of Isla Verde and on the north side from Isla Sacrificios. The depth of the profilers positioning was approximately 20 m, a vertical distribution of 2 m and a data log with time intervals of every 15 minutes, in addition to a vertical whitening distance of 0.75 m from the surface and 1 m in the bottom.
The ADCPs were used in the present work to measure the vertical profiles of the velocity components of the u and v component of the marine current, as well as the mean sea level, water temperature and the absolute acoustic intensity that later became intensity of backscatter or plankton biovolumes following the methodology of Salas-Monreal et al., (2009), where only the biovolumen of the north-south component of the plankton was considered, because it is the main direction of the Suradas and Nortes wind events.
Fluctuations in sea level and water temperature were measured by ADCPs of approximately at the depth of deployment (Salas-Pérez et al., 2012), which are the maximum depths of the reefs located in the northern area in the PNSAV.
Daily wind vectors, sea level atmospheric pressure and air temperature were extracted from the National Oceanic and Atmospheric Administration (NOAA) by NARR-NCEP (North American Regional Reanalysis-National Centers for Environmental Prediction): Data collected from November 10, 2008, to November 29, 2008 (Salas-Pérez et al., 2012). This is the period in which the presence of itinerant low- and -high pressure systems, passed for the study area. The oceanographic parameters, lees the biovolumen, were plotted with the original data filtered with a moving average filter (24 ,24 ,25) to remove the contribution of the diurnal tide in the study area (Salas-Pérez et al., 2008). Also, a linear trend was computed to observe the main pattern of the fluctuation during the Surada and Norte wind events.